feat(auth): refactor authentication routes to new user directory structure
